edgeServiceLib/object-recognition/object-recognition/requirements.txt
flask
gevent
pillow
requests
keras
uwsgi
tensorflow 